Fix Windows compile error.
Add _USE_MATH_DEFINES to the windows compile definitions.

+++ src/Core/CollisionPair.cpp
@@ -84,7 +84,7 @@ const std::vector< std::shared_ptr< Contact > >& CollisionPair::getContacts() co
 void CollisionPair::printCollisionPairs()
 {
     std::cout << "# Contacts: " << this->contacts.size() << std::endl;
-    for (int i = 0; i < this->contacts.size(); i++)
+    for (size_t i = 0; i < this->contacts.size(); i++)
     {
         std::cout << "Contact no: " << i << std::endl;
         this->contacts[0]->printInfo();
